requirements.txt
1.Streamlit
2.Matplotlib